Remove all script files
Since the tendency of this repository is to getting bigger over time,
given the increase of tweaks, customizations and programs that i've use,
and that this tendency also reflects the number os scripts that I wrote
and use, it makes sense to create a clear differentiation between
"dotfiles" and "scripts".

The scripts have been deleted and will be reallocated within the
"utility-belt" repository, also on my GitHub profile.