fix(rwal): issue where the templates cannot be copied
This commit fixes a problem where `rwal` cannot copy the templates from other
profiles (such as `bumblebee` or `rofi`) given the way GNU Stow make the
symlinks. Previously, `rwal` was trying to "guess" it's own absolute location
and, therefore, the location of the runners and templates. Since the templates
were located on other profile, the runners was failling in copy them.

This commit removes the "guessing" part and includes the absolute path of `rwal`.
It's important to notice that, although this path is absolute, it's not the
real one.